Make compendiums searchable
