+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/compendium/CompendiumScreen.kt
@@ -1,222 +1,100 @@
 package cz.frantisekmasa.wfrp_master.common.compendium
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyItemScope
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material.FloatingActionButton
 import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Scaffold
-import androidx.compose.material.Text
-import androidx.compose.material.TopAppBar
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Add
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.Dp
 import cafe.adriel.voyager.core.screen.Screen
-import cafe.adriel.voyager.navigator.LocalNavigator
-import cafe.adriel.voyager.navigator.currentOrThrow
-import cz.frantisekmasa.wfrp_master.common.compendium.blessing.BlessingCompendiumTab
-import cz.frantisekmasa.wfrp_master.common.compendium.career.CareerCompendiumTab
+import cz.frantisekmasa.wfrp_master.common.compendium.blessing.BlessingCompendiumScreen
+import cz.frantisekmasa.wfrp_master.common.compendium.career.CareerCompendiumScreen
 import cz.frantisekmasa.wfrp_master.common.compendium.domain.CompendiumItem
-import cz.frantisekmasa.wfrp_master.common.compendium.miracle.MiracleCompendiumTab
-import cz.frantisekmasa.wfrp_master.common.compendium.skill.SkillCompendiumTab
-import cz.frantisekmasa.wfrp_master.common.compendium.spell.SpellCompendiumTab
-import cz.frantisekmasa.wfrp_master.common.compendium.talent.TalentCompendiumTab
-import cz.frantisekmasa.wfrp_master.common.compendium.trait.TraitCompendiumTab
-import cz.frantisekmasa.wfrp_master.common.core.PartyScreenModel
+import cz.frantisekmasa.wfrp_master.common.compendium.miracle.MiracleCompendiumScreen
+import cz.frantisekmasa.wfrp_master.common.compendium.skill.SkillCompendiumScreen
+import cz.frantisekmasa.wfrp_master.common.compendium.spell.SpellCompendiumScreen
+import cz.frantisekmasa.wfrp_master.common.compendium.talent.TalentCompendiumScreen
+import cz.frantisekmasa.wfrp_master.common.compendium.trait.TraitCompendiumScreen
+import cz.frantisekmasa.wfrp_master.common.core.domain.NamedEnum
+import cz.frantisekmasa.wfrp_master.common.core.domain.localizedName
 import cz.frantisekmasa.wfrp_master.common.core.domain.party.PartyId
 import cz.frantisekmasa.wfrp_master.common.core.ui.buttons.BackButton
-import cz.frantisekmasa.wfrp_master.common.core.ui.dialogs.DialogState
 import cz.frantisekmasa.wfrp_master.common.core.ui.flow.collectWithLifecycle
-import cz.frantisekmasa.wfrp_master.common.core.ui.menu.DropdownMenuItem
 import cz.frantisekmasa.wfrp_master.common.core.ui.menu.WithContextMenu
 import cz.frantisekmasa.wfrp_master.common.core.ui.primitives.ContextMenu
-import cz.frantisekmasa.wfrp_master.common.core.ui.primitives.FullScreenProgress
-import cz.frantisekmasa.wfrp_master.common.core.ui.primitives.Spacing
-import cz.frantisekmasa.wfrp_master.common.core.ui.primitives.rememberScreenModel
-import cz.frantisekmasa.wfrp_master.common.core.ui.scaffolding.OptionsAction
-import cz.frantisekmasa.wfrp_master.common.core.ui.scaffolding.Subtitle
-import cz.frantisekmasa.wfrp_master.common.core.ui.scaffolding.tabs.TabPager
-import cz.frantisekmasa.wfrp_master.common.core.ui.scaffolding.tabs.tab
+import cz.frantisekmasa.wfrp_master.common.core.ui.primitives.SearchableList
 import cz.frantisekmasa.wfrp_master.common.localization.LocalStrings
+import cz.frantisekmasa.wfrp_master.common.localization.Strings
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
 import kotlin.coroutines.EmptyCoroutineContext
 
-class CompendiumScreen(
-    private val partyId: PartyId,
-) : Screen {
+abstract class CompendiumScreen() : Screen {
 
-    @Composable
-    override fun Content() {
-        val screenModel: PartyScreenModel = rememberScreenModel(arg = partyId)
-        Scaffold(topBar = { TopBar(screenModel) }) {
-            val strings = LocalStrings.current.compendium
-
-            TabPager(Modifier.fillMaxSize()) {
-                tab(strings.tabSkills) {
-                    SkillCompendiumTab(partyId, rememberScreenModel(arg = partyId), screenWidth)
-                }
-                tab(strings.tabTalents) {
-                    TalentCompendiumTab(partyId, rememberScreenModel(arg = partyId), screenWidth)
-                }
-                tab(strings.tabSpells) {
-                    SpellCompendiumTab(partyId, rememberScreenModel(arg = partyId), screenWidth)
-                }
-                tab(strings.tabBlessings) {
-                    BlessingCompendiumTab(partyId, rememberScreenModel(arg = partyId), screenWidth)
-                }
-                tab(strings.tabMiracles) {
-                    MiracleCompendiumTab(partyId, rememberScreenModel(arg = partyId), screenWidth)
-                }
-                tab(strings.tabTraits) {
-                    TraitCompendiumTab(partyId, rememberScreenModel(arg = partyId), screenWidth)
-                }
-                tab(strings.tabCareers) {
-                    CareerCompendiumTab(partyId, rememberScreenModel(arg = partyId), screenWidth)
-                }
-            }
-        }
+    enum class Type(
+        override val nameResolver: (strings: Strings) -> String,
+        val screen: (PartyId) -> Screen,
+    ) : NamedEnum {
+        SKILLS({ it.compendium.titleSkills }, ::SkillCompendiumScreen),
+        TALENTS({ it.compendium.titleTalents }, ::TalentCompendiumScreen),
+        SPELLS({ it.compendium.titleSpells }, ::SpellCompendiumScreen),
+        BLESSINGS({ it.compendium.titleBlessings }, ::BlessingCompendiumScreen),
+        MIRACLES({ it.compendium.titleMiracles }, ::MiracleCompendiumScreen),
+        TRAITS({ it.compendium.titleTraits }, ::TraitCompendiumScreen),
+        CAREERS({ it.compendium.titleCareers }, ::CareerCompendiumScreen),
     }
 
     @Composable
-    private fun TopBar(screenModel: PartyScreenModel) {
-        val strings = LocalStrings.current.compendium
-        val navigator = LocalNavigator.currentOrThrow
+    protected fun <T : CompendiumItem<T>> ItemsList(
+        liveItems: Flow<List<T>>,
+        emptyUI: @Composable () -> Unit,
+        onClick: (T) -> Unit,
+        remover: suspend (T) -> Unit,
+        newItemSaver: suspend (T) -> Unit,
+        onNewItemRequest: () -> Unit,
+        type: Type,
+        itemContent: @Composable LazyItemScope.(T) -> Unit,
+    ) {
+        val coroutineScope = rememberCoroutineScope { EmptyCoroutineContext + Dispatchers.IO }
 
-        TopAppBar(
-            title = {
-                Column {
-                    Text(strings.title)
-                    screenModel.party.collectWithLifecycle(null).value?.let {
-                        Subtitle(it.name)
-                    }
+        SearchableList(
+            data = liveItems.collectWithLifecycle(null).value
+                ?.let { SearchableList.Data.Loaded(it) } ?: SearchableList.Data.Loading,
+            emptyUi = emptyUI,
+            title = type.localizedName,
+            searchPlaceholder = "",
+            searchableValue = { it.name },
+            navigationIcon = { BackButton() },
+            key = { it.id },
+            itemContent = { item ->
+                WithContextMenu(
+                    items = listOf(
+                        ContextMenu.Item(
+                            LocalStrings.current.commonUi.buttonDuplicate,
+                            onClick = {
+                                coroutineScope.launch { newItemSaver(item.duplicate()) }
+                            }
+                        ),
+                        ContextMenu.Item(
+                            LocalStrings.current.commonUi.buttonRemove,
+                            onClick = { coroutineScope.launch { remover(item) } }
+                        ),
+                    ),
+                    onClick = { onClick(item) },
+                ) {
+                    itemContent(item)
                 }
             },
-            navigationIcon = { BackButton() },
-            actions = {
-                OptionsAction {
-                    DropdownMenuItem(
-                        onClick = { navigator.push(RulebookCompendiumImportScreen(partyId)) }
-                    ) {
-                        Text(strings.buttonImportFromRulebook)
-                    }
-
-                    DropdownMenuItem(
-                        onClick = { navigator.push(JsonCompendiumImportScreen(partyId)) }
-                    ) {
-                        Text(strings.buttonImportFile)
-                    }
-
-                    DropdownMenuItem(
-                        onClick = { navigator.push(JsonCompendiumExportScreen(partyId)) }
-                    ) {
-                        Text(strings.buttonExportFile)
-                    }
+            floatingActionButton = {
+                FloatingActionButton(onClick = onNewItemRequest) {
+                    Icon(
+                        Icons.Rounded.Add,
+                        LocalStrings.current.compendium.iconAddCompendiumItem,
+                    )
                 }
             }
         )
     }
-}
-
-@Composable
-fun <T : CompendiumItem<T>> CompendiumTab(
-    liveItems: Flow<List<T>>,
-    width: Dp,
-    emptyUI: @Composable () -> Unit,
-    onClick: (T) -> Unit,
-    remover: suspend (T) -> Unit,
-    newItemSaver: suspend (T) -> Unit,
-    onNewItemRequest: () -> Unit,
-    itemContent: @Composable LazyItemScope.(T) -> Unit,
-) {
-    Scaffold(
-        modifier = Modifier
-            .width(width)
-            .fillMaxHeight(),
-        floatingActionButton = {
-            FloatingActionButton(onClick = onNewItemRequest) {
-                Icon(
-                    Icons.Rounded.Add,
-                    LocalStrings.current.compendium.iconAddCompendiumItem,
-                )
-            }
-        }
-    ) {
-        Column(
-            Modifier
-                .fillMaxSize()
-                .background(MaterialTheme.colors.background)
-        ) {
-            val coroutineScope = rememberCoroutineScope { EmptyCoroutineContext + Dispatchers.IO }
-            val items = liveItems.collectWithLifecycle(null).value
-
-            when {
-                items == null -> FullScreenProgress()
-                items.isEmpty() -> emptyUI()
-                else -> {
-                    LazyColumn(contentPadding = PaddingValues(bottom = Spacing.bottomPaddingUnderFab)) {
-                        items(items, key = { it.id }) { item ->
-                            WithContextMenu(
-                                items = listOf(
-                                    ContextMenu.Item(
-                                        LocalStrings.current.commonUi.buttonDuplicate,
-                                        onClick = {
-                                            coroutineScope.launch { newItemSaver(item.duplicate()) }
-                                        }
-                                    ),
-                                    ContextMenu.Item(
-                                        LocalStrings.current.commonUi.buttonRemove,
-                                        onClick = { coroutineScope.launch { remover(item) } }
-                                    ),
-                                ),
-                                onClick = { onClick(item) },
-                            ) {
-                                itemContent(item)
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun <T : CompendiumItem<T>> CompendiumTab(
-    liveItems: Flow<List<T>>,
-    width: Dp,
-    emptyUI: @Composable () -> Unit,
-    dialog: @Composable (MutableState<DialogState<T?>>) -> Unit,
-    remover: suspend (T) -> Unit,
-    saver: suspend (T) -> Unit,
-    itemContent: @Composable LazyItemScope.(T) -> Unit,
-) {
-    val dialogState = remember { mutableStateOf<DialogState<T?>>(DialogState.Closed()) }
-
-    dialog(dialogState)
-
-    CompendiumTab(
-        liveItems = liveItems,
-        width = width,
-        emptyUI = emptyUI,
-        onClick = { dialogState.value = DialogState.Opened(it) },
-        onNewItemRequest = { dialogState.value = DialogState.Opened(null) },
-        remover = remover,
-        newItemSaver = saver,
-        itemContent = itemContent,
-    )
 }
