+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/localization/Strings.kt
@@ -528,13 +528,13 @@ data class CompendiumStrings(
     val pickPromptMiracles: String = "Select which miracles you want to import.",
     val pickPromptTraits: String = "Select which traits you want to import.",
     val rulebookStoreLink: String = "https://www.drivethrurpg.com/product/248284/Warhammer-Fantasy-Roleplay-Fourth-Edition-Rulebook?affiliate_id=2708720",
-    val tabBlessings: String = "Blessings",
-    val tabCareers: String = "Careers",
-    val tabMiracles: String = "Miracles",
-    val tabSkills: String = "Skills",
-    val tabSpells: String = "Spells",
-    val tabTalents: String = "Talents",
-    val tabTraits: String = "Traits",
+    val titleBlessings: String = "Blessings",
+    val titleCareers: String = "Careers",
+    val titleMiracles: String = "Miracles",
+    val titleSkills: String = "Skills",
+    val titleSpells: String = "Spells",
+    val titleTalents: String = "Talents",
+    val titleTraits: String = "Traits",
     val title: String = "Compendium",
     val titleImportCompendium: String = "Import Compendium",
     val titleExportCompendium: String = "Export Compendium",
